Move Tauri config into src-tauri
