+++ .gitignore
@@ -1,2 +1,3 @@
 __pycache__/
 logs/
+desktop_ui/node_modules/
