+++ core/automation_manager.py
