+++ desktop_ui/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "kari-control-room",
+  "version": "0.1.0",
+  "scripts": {
+    "dev": "npm --prefix frontend run dev",
+    "build": "npm --prefix frontend run build"
+  }
+}
