+++ desktop_ui/src-tauri/tauri.conf.json
@@ -1,22 +1,25 @@
 {
+  "package": {
+    "productName": "Kari Control Room",
+    "version": "0.1.0"
+  },
   "build": {
     "beforeBuildCommand": "npm run build",
     "beforeDevCommand": "npm run dev",
     "devPath": "http://localhost:3000",
     "distDir": "../frontend/dist"
   },
-  "package": {
-    "productName": "Kari Control Room",
-    "version": "0.1.0"
-  },
   "tauri": {
+    "bundle": {
+      "identifier": "com.example.kari"
+    },
+    "security": {
+      "csp": "default-src 'self'; connect-src http://localhost:8000"
+    },
     "allowlist": {
       "http": {
         "request": true
       }
-    },
-    "security": {
-      "csp": "default-src 'self'; connect-src http://localhost:8000"
     }
   }
 }
